Restructure project and add frontend React app
Moved backend files into a 'backend' directory and added a new React frontend in the 'frontend' directory. The backend Flask app now includes an endpoint to search presidents by date. On frontend, completed a table to view the data from the 'presidents' endpoint + a tab that allows you to search presidents by date.